[ao] Create framework for ModelReport Qconfig Generation
Summary: This creates the framework in the ModelReport API for the
generation of QConfigs by the ModelReport instance based on suggestions.
This functionality will eventually be added into the report generation
or be something that complements it, but for now it will be an
independent call for API stability and to be able to better modularize
the features as it stabilizes.

This also adds the framework for the relavent test function and a note
in the README at what future changes are planned for this new method in
the ModelReport API.

Test Plan: python test/test_quantization.py TestFxModelReportClass.test_qconfig_generation

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 968b0ca35cec98c8d33a8b20b5a18341ddff5627
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83091